Point lights working

+++ src/IDKgraphics/IDKgl/idk_gltools.hpp
@@ -9,7 +9,7 @@
 
 namespace idk::gltools
 {
-    GLuint  loadTexture( int w, int h, uint32_t *data, bool srgb = false );
+    GLuint  loadTexture( size_t w, size_t h, uint32_t *data, bool srgb = false );
     GLuint  loadTexture( std::string filepath, bool srgb = false );
 
 };
